user cython for fields.py, parse.py and validators.py
