+++ Makefile
@@ -6,7 +6,6 @@ black = black -S -l 120 --target-version py36 pydantic tests
 install:
 	pip install -U setuptools pip
 	pip install -U -r requirements.txt
-	pip install -e .
 
 .PHONY: format
 format:
@@ -26,6 +25,7 @@ mypy:
 
 .PHONY: test
 test:
+	pip install -e .
 	pytest --cov=pydantic
 
 .PHONY: external-mypy
@@ -43,6 +43,7 @@ external-mypy:
 
 .PHONY: testcov
 testcov:
+	pip install -e .
 	pytest --cov=pydantic
 	@echo "building coverage html"
 	@coverage html
