+++ tests/test_edge_cases.py
@@ -5,17 +5,7 @@ from typing import Any, Dict, List, Optional, Set, Tuple, Union
 
 import pytest
 
-from pydantic import (
-    BaseConfig,
-    BaseModel,
-    Extra,
-    NoneStrBytes,
-    StrBytes,
-    ValidationError,
-    constr,
-    errors,
-    validate_model,
-)
+from pydantic import BaseConfig, BaseModel, Extra, NoneStrBytes, StrBytes, ValidationError, constr, errors, parse_model
 
 
 def test_str_bytes():
@@ -601,12 +591,12 @@ def test_return_errors_ok():
         foo: int
         bar: List[int]
 
-    assert validate_model(Model, {'foo': '123', 'bar': (1, 2, 3)}) == (
+    assert parse_model(Model, {'foo': '123', 'bar': (1, 2, 3)}) == (
         {'foo': 123, 'bar': [1, 2, 3]},
         {'foo', 'bar'},
         None,
     )
-    d, f, e = validate_model(Model, {'foo': '123', 'bar': (1, 2, 3)}, False)
+    d, f, e = parse_model(Model, {'foo': '123', 'bar': (1, 2, 3)}, False)
     assert d == {'foo': 123, 'bar': [1, 2, 3]}
     assert f == {'foo', 'bar'}
     assert e is None
@@ -617,12 +607,12 @@ def test_return_errors_error():
         foo: int
         bar: List[int]
 
-    d, f, e = validate_model(Model, {'foo': '123', 'bar': (1, 2, 'x')}, False)
+    d, f, e = parse_model(Model, {'foo': '123', 'bar': (1, 2, 'x')}, False)
     assert d == {'foo': 123}
     assert f == {'foo', 'bar'}
     assert e.errors() == [{'loc': ('bar', 2), 'msg': 'value is not a valid integer', 'type': 'type_error.integer'}]
 
-    d, f, e = validate_model(Model, {'bar': (1, 2, 3)}, False)
+    d, f, e = parse_model(Model, {'bar': (1, 2, 3)}, False)
     assert d == {'bar': [1, 2, 3]}
     assert f == {'bar'}
     assert e.errors() == [{'loc': ('foo',), 'msg': 'field required', 'type': 'value_error.missing'}]
